+++ test/validation_message_test.js
@@ -1,8 +1,7 @@
 /* eslint-env mocha */
 const path = require('path')
 const assert = require('assert')
-const $rdf = require('rdf-ext')
-const rdfFS = require('rdf-utils-fs')
+const { loadDataset } = require('./utils')
 
 const SHACLValidator = require('../index')
 const rootPath = path.join(__dirname, '/data/validation-message')
@@ -10,7 +9,7 @@ const rootPath = path.join(__dirname, '/data/validation-message')
 describe('validation messages', () => {
   it('Returns message from shape property if provided', async () => {
     const dataPath = path.join(rootPath, 'message-from-shape-property.ttl')
-    const data = await $rdf.dataset().import(rdfFS.fromFile(dataPath))
+    const data = await loadDataset(dataPath)
     const shapes = data
 
     const validator = new SHACLValidator(shapes)
@@ -22,7 +21,7 @@ describe('validation messages', () => {
 
   it('Returns message from validator if provided (and no message on shape)', async () => {
     const dataPath = path.join(rootPath, 'message-from-validator.ttl')
-    const data = await $rdf.dataset().import(rdfFS.fromFile(dataPath))
+    const data = await loadDataset(dataPath)
     const shapes = data
 
     const validator = new SHACLValidator(shapes)
@@ -34,7 +33,7 @@ describe('validation messages', () => {
 
   it('Returns message from constraint component if provided (and no message on shape or validator)', async () => {
     const dataPath = path.join(rootPath, 'message-from-constraint-component.ttl')
-    const data = await $rdf.dataset().import(rdfFS.fromFile(dataPath))
+    const data = await loadDataset(dataPath)
     const shapes = data
 
     const validator = new SHACLValidator(shapes)
@@ -46,7 +45,7 @@ describe('validation messages', () => {
 
   it('Returns null if no message is defined anywhere', async () => {
     const dataPath = path.join(rootPath, 'message-empty.ttl')
-    const data = await $rdf.dataset().import(rdfFS.fromFile(dataPath))
+    const data = await loadDataset(dataPath)
     const shapes = data
 
     const validator = new SHACLValidator(shapes)
@@ -58,7 +57,7 @@ describe('validation messages', () => {
 
   it('Substitutes variables in message', async () => {
     const dataPath = path.join(rootPath, 'message-with-variable.ttl')
-    const data = await $rdf.dataset().import(rdfFS.fromFile(dataPath))
+    const data = await loadDataset(dataPath)
     const shapes = data
 
     const validator = new SHACLValidator(shapes)